Rename functions
General code tidy up
